♻️ refactor: query provider 내 로직 분리 #23
- getQueryClient / makeQueryClient 함수 분리
