+++ src/app/utils/query.ts
@@ -1,0 +1,26 @@
+import { QueryClient } from '@tanstack/react-query';
+
+let browserQueryClient: QueryClient | undefined;
+
+const makeQueryClient = () => {
+  return new QueryClient({
+    defaultOptions: {
+      queries: {
+        staleTime: 60 * 1000,
+        refetchOnMount: false,
+        refetchOnWindowFocus: false,
+      },
+    },
+  });
+};
+
+const getQueryClient = () => {
+  if (typeof window === 'undefined') {
+    return makeQueryClient();
+  }
+
+  if (!browserQueryClient) browserQueryClient = makeQueryClient();
+  return browserQueryClient;
+};
+
+export { makeQueryClient, getQueryClient };
